CSP-2440 Moved the new endpoint  under correct controller
