+++ src/SFA.DAS.RoATPService.Api.UnitTests/Controllers/OrganisationStatusEventsControllerTests.cs
@@ -1,21 +1,42 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
 using AutoFixture.NUnit3;
+using FluentAssertions;
 using MediatR;
+using Microsoft.AspNetCore.Mvc;
 using Moq;
 using NUnit.Framework;
 using SFA.DAS.RoATPService.Application.Api.Controllers;
 using SFA.DAS.RoATPService.Application.Queries.GetOrganisationsStatusEvents;
+using SFA.DAS.RoATPService.Application.Queries.GetOrganisationStatusHistory;
 using SFA.DAS.Testing.AutoFixture;
 
-namespace SFA.DAS.RoATPService.Api.UnitTests.Controllers.OrganisationsControllerTests;
+namespace SFA.DAS.RoATPService.Api.UnitTests.Controllers;
 
-public class OrganisationsControllerGetOrganisationStatusEventsTests
+public class OrganisationStatusEventsControllerTests
 {
     [Test, MoqAutoData]
-    public async Task GetOrganisationStatusEvents_InvokesMediator(
+    public async Task GetStatusEventsForUkprn_ReturnsOkResult(
         [Frozen] Mock<IMediator> mediatorMock,
-        [Greedy] OrganisationsController sut,
+        [Greedy] OrganisationStatusEventsController sut,
+        int ukprn,
+        GetOrganisationStatusHistoryQueryResult queryResult,
+        CancellationToken cancellationToken)
+    {
+        mediatorMock
+            .Setup(m => m.Send(It.Is<GetOrganisationStatusHistoryQuery>(q => q.Ukprn == ukprn), cancellationToken))
+            .ReturnsAsync(queryResult);
+
+        var result = await sut.GetStatusEventsForUkprn(ukprn, cancellationToken);
+
+        Assert.That(result, Is.TypeOf<OkObjectResult>());
+        result.As<OkObjectResult>().Value.Should().Be(queryResult);
+    }
+
+    [Test, MoqAutoData]
+    public async Task GetAllStatusEvents_InvokesMediator(
+        [Frozen] Mock<IMediator> mediatorMock,
+        [Greedy] OrganisationStatusEventsController sut,
         GetOrganisationStatusEventsQueryResult expected,
         int sinceEventId,
         int pageSize,
@@ -26,7 +47,7 @@ public class OrganisationsControllerGetOrganisationStatusEventsTests
             .Setup(m => m.Send(It.IsAny<GetOrganisationStatusEventsQuery>(), cancellationToken))
             .ReturnsAsync(expected);
 
-        await sut.GetOrganisationStatusEvents(sinceEventId, pageSize, pageNumber, cancellationToken);
+        await sut.GetAllStatusEvents(sinceEventId, pageSize, pageNumber, cancellationToken);
 
         mediatorMock.Verify(m => m.Send(
             It.Is<GetOrganisationStatusEventsQuery>(q =>
@@ -38,10 +59,10 @@ public class OrganisationsControllerGetOrganisationStatusEventsTests
 
     [MoqInlineAutoData(-1)]
     [MoqInlineAutoData(1001)]
-    public async Task GetOrganisationStatusEvents_DefaultsFilters_AndInvokesMediator(
+    public async Task GetAllStatusEvents_DefaultsFilters_AndInvokesMediator(
         int pageSize,
         [Frozen] Mock<IMediator> mediatorMock,
-        [Greedy] OrganisationsController sut,
+        [Greedy] OrganisationStatusEventsController sut,
         GetOrganisationStatusEventsQueryResult expected,
         CancellationToken cancellationToken)
     {
@@ -52,7 +73,7 @@ public class OrganisationsControllerGetOrganisationStatusEventsTests
             .Setup(m => m.Send(It.IsAny<GetOrganisationStatusEventsQuery>(), cancellationToken))
             .ReturnsAsync(expected);
 
-        await sut.GetOrganisationStatusEvents(sinceEventId, pageSize, pageNumber, cancellationToken);
+        await sut.GetAllStatusEvents(sinceEventId, pageSize, pageNumber, cancellationToken);
 
         mediatorMock.Verify(m => m.Send(
             It.Is<GetOrganisationStatusEventsQuery>(q =>
